feat(TimeStamp): update component
